feat: added proper data & links to dropdown navbar
