+++ app/[locale]/header.tsx
@@ -30,53 +30,88 @@ export default async function Header({ locale }: { locale: string }) {
   const isMacOS = agent.os.name === 'Mac OS';
 
   const items = [
-    { label: text.institute, href: 'institute' },
-    { label: text.academics, href: 'academics' },
     { label: text.faculty, href: 'faculty-and-staff' },
     { label: text.placement, href: 'training-and-placement' },
-    { label: text.alumni, href: 'alumni' },
+    // { label: text.alumni, href: 'alumni' },
     { label: text.activities, href: 'student-activities' },
   ];
-  const components: { title: string; href: string; description: string }[] = [
-    {
-      title: 'Alert Dialog',
-      href: '/docs/primitives/alert-dialog',
-      description:
-        'A modal dialog that interrupts the user with important content and expects a response.',
-    },
-    {
-      title: 'Hover Card',
-      href: '/docs/primitives/hover-card',
-      description:
-        'For sighted users to preview content available behind a link.',
-    },
-    {
-      title: 'Progress',
-      href: '/docs/primitives/progress',
-      description:
-        'Displays an indicator showing the completion progress of a task, typically displayed as a progress bar.',
-    },
-    {
-      title: 'Scroll-area',
-      href: '/docs/primitives/scroll-area',
-      description: 'Visually or semantically separates content.',
-    },
-    {
-      title: 'Tabs',
-      href: '/docs/primitives/tabs',
-      description:
-        'A set of layered sections of content—known as tab panels—that are displayed one at a time.',
-    },
-    {
-      title: 'Tooltip',
-      href: '/docs/primitives/tooltip',
-      description:
-        'A popup that displays information related to an element when the element receives keyboard focus or the mouse hovers over it.',
-    },
+
+  const dropDownItems = [
+    { label: text.institute, href: 'institute' },
+    { label: text.academics, href: 'academics' },
   ];
+  const academicItems: { title: string; href: string; description: string }[] =
+    [
+      {
+        title: 'Departments',
+        href: '/academics/departments',
+        description:
+          'Explore the diverse academic departments and their offerings.',
+      },
+      {
+        title: 'Programmes',
+        href: '/academics/programmes',
+        description:
+          'Discover our range of undergraduate and postgraduate programmes.',
+      },
+      {
+        title: 'Courses & Curricula',
+        href: '/academics/curricula',
+        description: 'Browse through the detailed list of courses available.',
+      },
+      {
+        title: 'Convocation',
+        href: '/academics/convocation',
+        description: 'Get information on upcoming convocation ceremonies.',
+      },
+      {
+        title: 'Awards',
+        href: '/academics/awards',
+        description:
+          'Recognizing excellence in academics, research, and beyond.',
+      },
+      {
+        title: 'Scholarship',
+        href: '/academics/scholarships',
+        description:
+          'Learn about scholarships, eligibility, and application details.',
+      },
+      {
+        title: 'Academic Notifications',
+        href: '/academics/notifications',
+        description:
+          'Stay updated with the latest academic announcements and deadlines.',
+      },
+    ];
+
+  const instituteItems: { title: string; href: string; description: string }[] =
+    [
+      {
+        title: 'Institute Profile',
+        href: '/institute/profile',
+        description: 'Get to know our institute’s vision, mission, and values.',
+      },
+      {
+        title: 'Administration',
+        href: '/institute/administration',
+        description: 'Meet the leadership team guiding our institution.',
+      },
+      {
+        title: 'Sections',
+        href: '/institute/sections',
+        description:
+          'Explore the various sections that support campus life and academics.',
+      },
+      {
+        title: 'Campus Infrastructure',
+        href: '/institute/campus-infra',
+        description:
+          'Discover the state-of-the-art facilities and infrastructure on campus.',
+      },
+    ];
 
   return (
-    <header className="header-sticky-ness sticky top-0 z-nav min-w-full bg-background">
+    <header className="header-sticky-ness sticky top-0 z-50 min-w-full bg-background">
       <nav
         className={cn(
           'container flex justify-between',
@@ -97,7 +132,22 @@ export default async function Header({ locale }: { locale: string }) {
           <NavigationMenuList
             className={cn('hidden grow lg:flex', 'gap-4 xl:gap-5 2xl:gap-6')}
           >
-            <NavigationMenuCustomListItem
+            {dropDownItems.map(({ label, href }, index) => (
+              <NavigationMenuCustomListItem
+                key={index}
+                triggerName={label}
+                imageDetails={{
+                  src: `${href}/image01.jpg`,
+                  alt: label,
+                  href: href,
+                }}
+                listItems={
+                  label === text.institute ? instituteItems : academicItems
+                }
+              />
+            ))}
+
+            {/* <NavigationMenuCustomListItem
               triggerName={items[0].label}
               imageDetails={{
                 src: 'https://s3-alpha-sig.figma.com/img/054e/19b7/43c945f2ee30e43f797f944b1c02fe2e?Expires=1728259200&Key-Pair-Id=APKAQ4GOSFWCVNEHN3O4&Signature=TjH6LvDVUoTwEvUvc02-8DAwccrJ9YRqDqTy5h-0O0cYSVWG03it8-zr5OSBcjGVuu5TMnV7ZlnEiM3CIHQ3mQAy7Z3~Bv2sbCL8plMbE0GDxzmjpaVkKPAfgbMpuyofWABnyQjH4cda6qWEzeBuGEw~KQfFxVuAA-wHYTA6GL~B776fRbfdfzNxtSqucrIEqfGG1nUMFEdxvTLMPCqXTjErPikIs2rDXtAZ3K3U4suPFFqLRyBQ9H0B3DAGDzxZ64CIVLkAaE~ALCRy1BBUDyXrU24E1~BeTobiCoR0q1WcnBOPMKpnUb0c2qTyaDz8BxMe9hMMI9vGnv4fxdqGBA__',
@@ -105,14 +155,23 @@ export default async function Header({ locale }: { locale: string }) {
                 href: items[0].href,
               }}
               listItems={components}
-            />
-            <NavigationMenuItem>
+            /> */}
+            {items.map(({ label, href }, index) => (
+              <NavigationMenuItem key={index}>
+                <Link href={`/${locale}/${href}`} legacyBehavior passHref>
+                  <NavigationMenuLink className={navigationMenuTriggerStyle()}>
+                    {label}
+                  </NavigationMenuLink>
+                </Link>
+              </NavigationMenuItem>
+            ))}
+            {/* <NavigationMenuItem>
               <Link href="/docs" legacyBehavior passHref>
                 <NavigationMenuLink className={navigationMenuTriggerStyle()}>
                   Documentation
                 </NavigationMenuLink>
               </Link>
-            </NavigationMenuItem>
+            </NavigationMenuItem> */}
           </NavigationMenuList>
         </NavigationMenu>
